Stupendous serialized session
This is the most daunting data structure yet! Almost every file in the
project was touched to implement this change, including rewriting a
heaping of shell scripts in C. This replaces the previous "session
directory" proof-of-concept and its various hacks.

The new session implementation introduces three main data structures:
step, input and dependency.

A session step corresponds to a plan step. It proceeds through an
abstract state machine where it is blocked on dependencies, available to
be resolved, then prepared with a job, and finally recorded with a
production. However if any of the dependencies it is blocked on are
missing, the step will become unresolvable. Note this terminology is
subject to refinement.

A number of inputs are associated with each step. When a step is
resolved to a job, its inputs will be used to associate input paths with
corresponding resources.

A dependency may additionally be associated with an unpopulated input,
specifying which step's output path (and corresponding resource) the
input should eventually be populated with; thus a preceding step's
output satisfies a succeeding step's input.

A step is blocked as long as it has pending dependencies (the upstream
steps have not finished). For a completed step to efficiently satisfy
its dependencies, the dependency data structures must be sorted by the
corresponding completing step and output path. Satisfying dependencies
involves iterating over both production outputs and a corresponding
range of dependencies to copy matching production output resources into
the associated dependency input.

Each data structure (except dependency) uses flags in a similar way that
is packed with the length of its corresponding name (step _name_, input
_path_, and dependency _output_).

A particular session is active (i.e., global variables set) according to
its filename in $KNIT_DIR/sessions. If no session is loaded, the
initially empty one may be populated and saved to create a new session.
To accommodate modifying loaded sessions, we make its mmap writable but
private; changes are saved through a new renamed session file.

Because knit-list-steps already displays information from the session,
it has been extended as a diagnostic tool with the --debug flag.

Also, introduce the concept of a stepish, which is either a step name or
step position of the form @<pos>. The latter should be preferred in
scripts as it is more direct (avoids a linear scan) and less ambiguous.