Partial steps implemented with linked lists
Some of the logic for handling partials already existed, but it wasn't
fully implemented. This ties those loose ends; in particular we copy
partials when they are instantiated by other steps, and we merge in any
additional inputs (overriding those from the partial). This also
reimplements inputs (and steps) as sorted linked lists.