Represent process as a special process input
Rather than parsing a process type that propagates into the session,
just add special inputs .knit/shell, .knit/param, and .knit/flow. The
flow process type needs a value anyway, and the plan is to replace shell
with cmd which would also take a value; having a special input is a
straightforward way to pass these values. knit-dispatch-job now looks
for the process input to determine how to execute (but still only
supports shell for now).

Another approach would be to still tag by process type and propagate it
all the way to the job object. The values for flow and cmd could be
specified through a "sentinel" input (probably named the empty string).
This is a little more prescriptive, but also requires wiring process
types through the parse tree, sessions, and jobs. From a data encoding
perspective, it would save a few bytes on shorter input names and
completely omit the unused input for param and shell.