Add cmd process type and knit-exec-cmd helper
The cmd process type is intended to replace shell. Its process input
specifies the command to run with NUL-delimited arguments.

To accommodate this we minimally extend parsing to support NUL bytes in
string literals. The escape sequence is the C-style '\000' for forward
compatibility; we don't support any other octal values, but it can be
shortened to '\0' where unambiguous. Technically lex_string() need not
append a NUL byte anymore, but we leave that behavior for now since we
would also need to change the semantics of lex_string_alloc().

knit-exec-cmd parses the cmd process input and executes it with some
setup, namely the working directory and closing or redirecting file
descriptors except stdout.