Create a flow job in knit-parse-plan
Encapsulate a flow plan into a job to allow future improvements. This
should become the basis for running subflows. It will also allow us to
cache entire flows, and it provides a distinct name for our session.

To serialize the job we add a function store_job() and rename the
previous session_store_job(); these functions are nearly identical but
one serializes resource_list and the other session_input.