Adjust knit-parse-plan options
Change param options to -p for literal and -P for file; repurpose the -f
option to specify the flow plan file; add a -F option to specify the
directory to search for files referenced in the flow plan.

Move populating of input resources before printing; we already have to
traverse inputs to check for holes, and it keeps printing simple as we
add complexity to specifying inputs.