Make flow job an explicit intermediate
knit-parse-plan is now invoked twice. The overall effect is still to
create a flow job and emit build instructions to knit-build-session.
However we now make processing before and after the flow job more
discrete. This should help to implement subflows later at the cost of
parsing the plan twice.

To create the flow job, we delegate to knit-plan-job. Some information
is needed from the parsed plan to create the flow job, namely files
referenced by the plan; we also provide any plan params for validation,
since knit-plan-job has also taken responsibility for parsing param
flags. These files and params are provided with the --emit-params-files
option to knit-parse-plan; this is a pure parse that reads only the plan
file. knit-plan-job will then actually read the referenced files and
output them in a flow job as inputs prefixed by files/, along with param
inputs prefixed by params/, and a .knit/flow plan.

knit-parse-plan can read a flow job with --job-to-session to output
build instructions as before. This will not access any files (outside of
flow job resources).

Example output of knit-parse-plan:

    $ cat plan.knit
    step params: params
        a = ./filea
        b = "x"
    step foo: identity
        b = params:b
        c = ./filec

    $ ./knit-parse-plan --emit-params-files plan.knit
    param a=filea
    param b
    file filec