Fix memory corruption in object trie
There were two errors in the object trie. One was out of bounds indexing
due to not masking the result of trie_child(). The other was not
reinitializing the trie node pointers to NULL when converting a leaf
node to internal.

The previous commit made these failures more likely because more objects
are interned. The segfault was found in the trie code, and the bad
access was further investigated with Valgrind and vgdb.

The object_trie fields were renamed to the less ambiguous items and nodes.