Add knit-peel-spec for parsing object ids
This is inspired by git-rev-parse (git help revisions). The intention is
to make it convenient to specify related object ids without
copy-and-pasting or verbose shell scripting. For example, we might
unwrap a job flow invocation with ^{production}^{invocation}.

The syntax is quite experimental, so we'll see how it shakes out. When
it's a bit firmer, we'll want to use this instead of hash_to_oid (which
incidentally is broken) to parse command line arguments.