Directory dependencies
To have a dependency on an entire directory of outputs, we rely on a
special input and a special dependency. The dependency has the SD_PREFIX
flag which indicates we only need a prefix (including the directory's
trailing '/') of the dependency output path to match the production
output.

This in turn sets the SI_FANOUT flag on the associated input, which
stores a fanout step position. This fanout step is not a real step, but
we create inputs for it for each matching output. The fanout step is
essentially a way to append new session inputs to an existing step
without having to insert and renumber existing inputs.

The full input name in case of fanout will be the parent SI_FANOUT input
name concatenated with the child input name (which is the production
output with the matching prefix removed). This gives the intuitive
result like copying from the output directory to the input directory.

Fanout steps introduce a non-determinism to sessions, since the order
they are added depends on when steps finish and dependencies resolved.
We may want to implement flow job caching in knit-run-plan to improve
reproducibility.