Add implicit dependency on .knit/ok by default
Most dependencies are only intended to resolve if their jobs run with a
successful exit code of 0. Such a production has a .knit/ok special
output. A typical dependency will now implicitly depend on this output
in addition to the declared dependency.

We also want to be able to declare a dependency that does not have such
a so-called "implicit ok" dependency, so we introduce a := syntax
(think, exact dependency) that suppresses the implicit dependency.

A session_dependency normally resolves through a session_input, but
that's unnecessary for an implicit ok dependency; it ignores the actual
result, so the resource and input name are unused. In this case, we
overload the input_pos field to directly contain the step position when
SD_INPUTISSTEP is set and skip having a session_input.