Parse @{step} in spec
The braced value may be a step name or non-zero numeric position; a
negative numeric value is interpreted from the last step. That is, @{-1}
is the terminal step.

A previous consideration has been to relax step name identifiers so they
need not be ASCII alphanumeric nor start with a letter; however some
care would be needed as that would make this spec syntax ambiguous (both
for numeric positions and brace delimiters).

Tangentially, we might consider allowing a bare : to coerce a production
to a job. This would allow a spec like @:: to get the terminal job.