Directory file inputs
If an input file is indicated as a directory by a trailing /, it will be
expanded to include each file inside that directory recursively. However
any top-level .knit directory will be ignored, so that ./ will exclude
any knit repository; it may be sensible to generalize this to
non-top-level .knit directories also.

Directory files may also be marked optional with ?= (like dependencies);
otherwise an empty directory will result in an error.

We now validate that inputs are disjoint. It's possible this could be
relaxed with rules for precedence.

knit-parse-plan --emit-params-files output has been adjusted slightly to
allow optional files and to prepend filenames with ./ (to make it
clearer when the entire plan directory is included).

The main changes are to knit-plan-job, so it includes all files inside
referenced directories in the flow job; and to knit-parse-plan so it
expands directory step inputs with any corresponding flow job inputs.