Environment variable input names
Allow input names to start with $ to indicate they will set an
environment variable. knit-unpack-job writes to an environ file that is
loaded by knit-exec-cmd.

In practice this is hard to use because we do not trim trailing newlines
(unlike shell "$(process substitution)"). Perhaps a syntax for newline
trimming would be appropriate.

This does not add any support for environment variables in input values.
However that would be interesting to add, since the environment is a
implicit dependency that currently leaks into jobs. We might even
sanitize the environment so all variables are explicit dependencies
(maybe with some exceptions; see the ENVIRONMENT section of man sudo).