Serialize invocation entries in reverse session order
Notably, the terminal step is now the first entry. This is sensible
because we implicitly dereference the invocation to its terminal step.
It also saves us a pointer in the parsed struct invocation.

We plan to also adjust spec peeling since in practice we usually want to
reference invocation entries relative to the terminal.