More ergonomic flow process syntax
Previously a flow process specified a flow plan as its value and its
inputs were expected to have special prefixes params/ and files/. This
is cumbersome for the common case of specifying a directory as a
subflow.

Instead a flow process now also takes a context, which is the directory
where plan files will be read from. Params are specified as inputs
without a prefix. The flow plan can be specified as before, or it can
use a new syntax :suffix.knit that will be evaluated relative to the
flow context; the plan may also be omitted to default to :plan.knit.

This is implemented by reading the context to inputs prefixed by
.knit/files/ and marking the step with is_raw_flow. In
rewrite_raw_flow_inputs(), we transform this to a files/ prefix and
other inputs to a params/ prefix to provide the same flow job format as
before. However we should consider simplifying this by removing this
transformation and reading params from unprefixed inputs and plan files
from .knit/files/.

We also fix propagation from a nocache partial. A future refactoring may
reorganize commonality between files and dependencies in struct value.