Simplify spec peeling
In practice it's simpler to use a single sigil for dereferencing
productions and invocations. We choose ~ since git has set a precedent
for that as a sigil. Peeling a chain of invocations has not yet proven
useful and has been removed. A bare ~ will dereference to an invocation,
while a braced ~{step} or numeric ~1 will dereference to a corresponding
production. This is somewhat similar to how a bare : will dereference to
a production or job, but with a :path it will deference to a resource.

Also fix invocation parsing to NUL-terminate step names and emit more
error messages.