Remove unused job from invocation entry
Originally job was included as well as production because we thought
there might be situations where a job would be present without a
production; that appears to not be the case. We can continue to retrieve
the job through the production.

A nicety of this is that invocation entries now fit on a single standard
line (for short step names).