Port knit-resume-session to C
Besides a slight performance improvement, the main reason for the port
is to ease implementation of more complex behavior. Although I now find
myself at a loss as to what that more complex behavior would be.

C is more cumbersome than bash at interfacing with child process. Some
of this should be smoothed out as dependencies are also ported to C and
we can call native functions.

Warnings about unmet requirements have been moved to knit-run-plan
(which is still written in bash). These will now display even when the
flow job is cached.