Fix PATH modification
We should only modify PATH if the knit command is called with an
absolute or relative path. Otherwise we would incorrectly add the
current working directory to PATH.

In the future we may replace this all with a compile-time KNIT_EXEC_PATH
similar to git. This could especially be the case if we install
subcommands to /usr/libexec.