Significant restructure of knit-schedule-jobs
The most important change is to split job_extra from dispatch. This
allows us to keep the notify_list for job completions separate from the
execution state of a job subprocess. This distinction will become more
important when we introduce external (or "manual") jobs which do not
have subprocesses.

Other changes:
- Restructure poll loop for readability and to avoid deletion during
  iteration.
- Split dispatch_session for session-specific dispatch state.
- Fix to update references to struct pollfd on deletion.
- Introduce DS_CACHE state to simplify reading from cache.