External (or manual) job process
Knit has long conceived of supporting manual jobs, which are not
executed as part of a flow, but instead their output are provided out of
band. External jobs (along with persistent sessions) finally make that
feasible. We can invoke flows that include manual steps and resume their
sessions even if knit is terminated; this could be thought of as
asynchronous operation. In practice this workflow is still unrefined.

knit-schedule-jobs now writes external jobs to stdout and reads their
completed productions on stdin; this means the read_buffer for stdin has
no associated dispatch. This protocol is analogous to the one for
requesting jobs used with knit-resume-session, but we prepend the word
"external" to disambiguate jobs from the final production prepended with
the word "ok". For example:

    external 6b13d94c405be04e95534c678d2aa937d521d31df288bb7aaf5b5b80bb5cc428
    ok 241c96a14db07d4d21946c49fb6c09c86d7acf52e0320baf62e4379c7a653710

where the first line is an external job (its production is provided to
stdin) and the second is the flow production.

knit-run-plan does not expose the external job protocol. Instead it uses
a named pipe at .knit/run.pipe to communicate completed productions.
knit-complete-job has been repurposed for this task (it has been unused
since it was previously used to write to sessions).

The named pipe is quite awkward. Perhaps we should consider a socket.