Replace .knit/run.pipe with .knit/run.sock
Expand the role of knit-filter-status to effectively become a server. It
listens on .knit/run.sock to service connections and mediates providing
productions to knit-schedule-jobs. Compared to the named pipe this
should provide a less finicky protocol for interprocess coordination.

It is easy to see this role growing further (receiving external job
requests, querying progress, dispatching flows) or consolidating with
knit-schedule-jobs.

Because listening on the socket is essential, disabling filtering with
--off now launches knit-schedule-jobs as a subprocess rather than
exec'ing it.