knit-build-objects to efficiently create knit objects
The precipitating use case is to plan jobs without having to create a
directory tree for the flow job.

We implement a simple stack-based interpreter so we can build dependent
objects (like a production of a job with resources) in a single process.

For example, to create an identity job with two inputs, and create and
output a production:
    resource 0
    entry .knit/identity
    resource 8
    input 1
    entry name1
    resource 8
    input 2
    entry name2
    link 2
    job
    resource 7
    output
    entry name
    production
    pop production

Also make job and production serialization more strict.

Notably absent is support for invocations (building them or including
them in productions).